CBG-2031 [3.0.1 backport] Fix _limit handling for user endpoint (#5522)
* CBG-2031 [3.0.1 backport] Fix _limit handling for user endpoint

Backports CBG-2030 to 3.0.1.

* Apply suggestions from code review
