+++ db/query.go
@@ -150,8 +150,15 @@ var QueryPrincipals = SGQuery{
 			"WHERE META(`%s`).id LIKE '%s' "+
 			"AND (META(`%s`).id LIKE '%s' "+
 			"OR META(`%s`).id LIKE '%s') "+
+			"AND META(`%s`).id >= $%s "+ // Using >= to match views inclusive startKey handling
 			"ORDER BY META(`%s`).id",
-		base.KeyspaceQueryToken, base.KeyspaceQueryToken, base.KeyspaceQueryToken, SyncDocWildcard, base.KeyspaceQueryToken, `\\_sync:user:%`, base.KeyspaceQueryToken, `\\_sync:role:%`, base.KeyspaceQueryToken),
+		base.KeyspaceQueryToken,
+		base.KeyspaceQueryToken,
+		base.KeyspaceQueryToken, SyncDocWildcard,
+		base.KeyspaceQueryToken, `\\_sync:user:%`,
+		base.KeyspaceQueryToken, `\\_sync:role:%`,
+		base.KeyspaceQueryToken, QueryParamStartKey,
+		base.KeyspaceQueryToken),
 	adhoc: false,
 }
 
@@ -441,19 +448,15 @@ func (context *DatabaseContext) QueryPrincipals(startKey string, limit int) (sgb
 
 	queryStatement := replaceIndexTokensQuery(QueryPrincipals.statement, sgIndexes[IndexSyncDocs], context.UseXattrs())
 
+	params := make(map[string]interface{})
+	params[QueryParamStartKey] = startKey
+
 	if limit > 0 {
 		queryStatement = fmt.Sprintf("%s LIMIT %d", queryStatement, limit)
 	}
 
-	params := make(map[string]interface{})
-	if startKey != "" {
-		queryStatement = fmt.Sprintf("%s AND META(`%s`).id >= $startkey",
-			queryStatement, context.Bucket.GetName())
-		params[QueryParamStartKey] = startKey
-	}
-
 	// N1QL Query
-	return context.N1QLQueryWithStats(QueryTypePrincipals, queryStatement, nil, base.RequestPlus, QueryPrincipals.adhoc)
+	return context.N1QLQueryWithStats(QueryTypePrincipals, queryStatement, params, base.RequestPlus, QueryPrincipals.adhoc)
 }
 
 // Query to retrieve the set of user and role doc ids, using the primary index
